test(cli): extract production play tests
Move BUILD unit/production command coverage, production chooser closeout verification, and production input validation into a focused play test file. Keep the monolith responsible for production-choice HUD projection, ready-city read-side candidates, notification queues, and compact priorities.

Include the focused file in test:cli:play so the canonical play suite continues to run the extracted coverage. Remove the broad fixture production send/postcondition state now that the focused file owns those command assertions.

Proof boundary: test ownership modularization only. No CLI runtime source, direct-control behavior, gameplay send, live-game proof, play-agent workflow, Studio transport, or oRPC ancestry is changed.
